requirements.txt
absl-py==1.0.0
beautifulsoup4==4.10.0
blessings==1.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
config-argparse==0.3.3
cycler==0.11.0
filelock==3.4.2
fonttools==4.28.5
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.43.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
kiwisolver==1.3.2
lpips==0.1.4
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.5.62
packaging==21.3
Pillow==8.4.0
plyfile==0.7.4
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.2.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3.
setuptools==59.5.0
six==1.16.0
soupsieve==2.3.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
warmup-scheduler==0.3
Werkzeug==2.0.2
zipp==3.7.0
